xfs_repair: add printf format checking and fix the fallout
Add the gcc printf like attribute to the xfs_repair-internal logging
helpers, and fix the massive fallout.  A large part of it is dealing
with the correct format for fixed size 64-bit types, but there were
a lot of real bug in there, including some that lead to crashed when
repairing certain corrupted filesystems on ARM based systems.

[Moved in a few more warning fixes from the next patch. -Alex]

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reported-by: Anisse Astier <anisse@astier.eu>
Signed-off-by: Alex Elder <aelder@sgi.com>