xfsprogs: xfs_quota: don't print invalid quota file inode number
When the state of quota files is dumped, xfs_quota blindly shows
whatever inode number is returned by the kernel.  If one of the
quota types is not enabled or enforced, the inode number provided
is an invalid value ((__u64) -1).  Rather than print a meaningless
large integer, print "N/A" in its place to make interpreting the
result it a little more obvious.

Signed-off-by: Alex Elder <aelder@sgi.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>