xfsprogs: document all commands in xfs_io
Add missing command documentation to xfs_io(8) manpage.
fiemap, fpunch, chproj, lsproj, and setfl are all missing.

setfl seems to not work today in any case, and nothing
in xfstests uses it; I will send another patch to simply
remove it from xfs_io, as I don't think it's terribly useful,
and hasn't worked forever anyway.

Also fix references to the fallocate manpage, which is (now?)
in section 2, not section 3 of the man pages.  (Since it's
a syscall, not a library function).

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>